requirements.txt
Flask~=2.2.2
gunicorn~=20.1.0
requests
python-whois
python-dotenv~=0.21.1
pip~=23.0.1
attrs~=22.2.0
cryptography~=36.0.1
auth~=0.5.3
filelock~=3.9.0
MarkupSafe~=2.1.2
Werkzeug~=2.2.2
click~=8.1.3
Jinja2~=3.0.3
blinker~=1.5
itsdangerous~=2.1.2
python-dateutil~=2.8.2
certifi~=2022.12.7
urllib3~=1.26.14
six~=1.16.0
packaging~=23.0
eventlet~=0.33.3
greenlet~=1.1.1
idna~=3.4
chardet~=5.1.0